Move actions to separate file
